+++ simpletransformers/question_answering/question_answering_model.py
@@ -233,6 +233,33 @@ class QuestionAnsweringModel:
         ):
             features = torch.load(cached_features_file)
             logger.info(f" Features loaded from cache at {cached_features_file}")
+
+            # Convert to Tensors and build dataset
+            all_input_ids = torch.tensor([f.input_ids for f in features], dtype=torch.long)
+            all_attention_masks = torch.tensor([f.attention_mask for f in features], dtype=torch.long)
+            all_token_type_ids = torch.tensor([f.token_type_ids for f in features], dtype=torch.long)
+            all_cls_index = torch.tensor([f.cls_index for f in features], dtype=torch.long)
+            all_p_mask = torch.tensor([f.p_mask for f in features], dtype=torch.float)
+            all_is_impossible = torch.tensor([f.is_impossible for f in features], dtype=torch.float)
+
+            if mode == "dev":
+                all_feature_index = torch.arange(all_input_ids.size(0), dtype=torch.long)
+                dataset = TensorDataset(
+                    all_input_ids, all_attention_masks, all_token_type_ids, all_feature_index, all_cls_index, all_p_mask
+                )
+            else:
+                all_start_positions = torch.tensor([f.start_position for f in features], dtype=torch.long)
+                all_end_positions = torch.tensor([f.end_position for f in features], dtype=torch.long)
+                dataset = TensorDataset(
+                    all_input_ids,
+                    all_attention_masks,
+                    all_token_type_ids,
+                    all_start_positions,
+                    all_end_positions,
+                    all_cls_index,
+                    all_p_mask,
+                    all_is_impossible,
+                )
         else:
             logger.info(" Converting to features started.")